// src/types/colors.ts
import { SimplePaletteColorOptions } from '@mui/material';

export type Color = {
  primary: SimplePaletteColorOptions;
  secondary: SimplePaletteColorOptions;
  error: SimplePaletteColorOptions;
  success: SimplePaletteColorOptions;
  textWhite: SimplePaletteColorOptions;
  textBlack: SimplePaletteColorOptions;
  paper: SimplePaletteColorOptions;
  border: SimplePaletteColorOptions;
};